Move client-to-server packets to c2s package
